+++ tests/beta/interfaces/test_tape_autograd.py
@@ -16,7 +16,7 @@ import pytest
 from pennylane import numpy as np
 
 import pennylane as qml
-from pennylane.beta.tapes import QuantumTape, qnode, QNode
+from pennylane.beta.tapes import QuantumTape
 from pennylane.beta.queuing import expval, var, sample, probs
 from pennylane.beta.interfaces.autograd import AutogradInterface
 
@@ -730,591 +730,3 @@ class TestAutogradPassthru:
         x = np.array(0.543, requires_grad=True)
         res = cost(x, device=dev)
         assert res.shape == (2, 10)
-
-
-@pytest.mark.parametrize(
-    "dev_name,diff_method",
-    [["default.qubit", "finite-diff"], ["default.qubit.autograd", "backprop"]],
-)
-class TestQNode:
-    """Same tests as above, but this time via the QNode interface!"""
-
-    def test_execution_no_interface(self, dev_name, diff_method):
-        """Test execution works without an interface"""
-        if diff_method == "backprop":
-            pytest.skip("Test does not support backprop")
-
-        dev = qml.device(dev_name, wires=1)
-
-        @qnode(dev, interface=None)
-        def circuit(a):
-            qml.RY(a, wires=0)
-            qml.RX(0.2, wires=0)
-            return expval(qml.PauliZ(0))
-
-        a = np.array(0.1, requires_grad=True)
-
-        res = circuit(a)
-
-        assert circuit.qtape.interface == None
-
-        # without the interface, the tape simply returns an array of results
-        assert isinstance(res, np.ndarray)
-        assert res.shape == (1,)
-        assert isinstance(res[0], float)
-
-        # without the interface, the tape is unable to deduce
-        # trainable parameters
-        assert circuit.qtape.trainable_params == {0, 1}
-
-        # gradients should cause an error
-        with pytest.raises(TypeError, match="must be real number, not ArrayBox"):
-            qml.grad(circuit)(a)
-
-    def test_execution_with_interface(self, dev_name, diff_method):
-        """Test execution works with the interface"""
-        if diff_method == "backprop":
-            pytest.skip("Test does not support backprop")
-
-        dev = qml.device(dev_name, wires=1)
-
-        @qnode(dev, interface="autograd")
-        def circuit(a):
-            qml.RY(a, wires=0)
-            qml.RX(np.array(0.2, requires_grad=False), wires=0)
-            return expval(qml.PauliZ(0))
-
-        a = np.array(0.1, requires_grad=True)
-        circuit(a)
-
-        assert circuit.qtape.interface == "autograd"
-
-        # the tape is able to deduce trainable parameters
-        assert circuit.qtape.trainable_params == {0}
-
-        # gradients should work
-        grad = qml.grad(circuit)(a)
-        assert len(grad) == 1
-        assert isinstance(grad[0], np.ndarray)
-        assert grad[0].shape == tuple()
-
-    def test_interface_swap(self, dev_name, diff_method, tol):
-        """Test that the autograd interface can be applied to a QNode
-        with a pre-existing interface"""
-        tf = pytest.importorskip("tensorflow", minversion="2.1")
-
-        if diff_method == "backprop":
-            pytest.skip("Test does not support backprop")
-
-        dev = qml.device(dev_name, wires=1)
-
-        @qnode(dev, interface="tf")
-        def circuit(a):
-            qml.RY(a, wires=0)
-            qml.RX(np.array(0.2, requires_grad=False), wires=0)
-            return expval(qml.PauliZ(0))
-
-        a = tf.Variable(0.1, dtype=tf.float64)
-
-        with tf.GradientTape() as tape:
-            res_tf = circuit(a)
-
-        grad_tf = tape.gradient(res_tf, a)
-
-        # switch to autograd interface
-        circuit.to_autograd()
-
-        a = np.array(0.1, requires_grad=True)
-
-        res = circuit(a)
-        grad = qml.grad(circuit)(a)
-        assert len(grad) == 1
-
-        assert np.allclose(res, res_tf, atol=tol, rtol=0)
-        assert np.allclose(grad[0], grad_tf, atol=tol, rtol=0)
-
-    def test_jacobian(self, dev_name, diff_method, mocker, tol):
-        """Test jacobian calculation"""
-        spy = mocker.spy(QuantumTape, "jacobian")
-        a = np.array(0.1, requires_grad=True)
-        b = np.array(0.2, requires_grad=True)
-
-        dev = qml.device(dev_name, wires=2)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(a, b):
-            qml.RY(a, wires=0)
-            qml.RX(b, wires=1)
-            qml.CNOT(wires=[0, 1])
-            return [expval(qml.PauliZ(0)), expval(qml.PauliY(1))]
-
-        res = circuit(a, b)
-
-        assert circuit.qtape.trainable_params == {0, 1}
-        assert res.shape == (2,)
-
-        expected = [np.cos(a), -np.cos(a) * np.sin(b)]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        res = qml.jacobian(circuit)(a, b)
-        expected = [[-np.sin(a), 0], [np.sin(a) * np.sin(b), -np.cos(a) * np.cos(b)]]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        if diff_method == "finite-diff":
-            spy.assert_called()
-        elif diff_method == "backprop":
-            spy.assert_not_called()
-
-    def test_jacobian_no_evaluate(self, dev_name, diff_method, mocker, tol):
-        """Test jacobian calculation when no prior circuit evaluation has been performed"""
-        spy = mocker.spy(QuantumTape, "jacobian")
-        a = np.array(0.1, requires_grad=True)
-        b = np.array(0.2, requires_grad=True)
-
-        dev = qml.device(dev_name, wires=2)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(a, b):
-            qml.RY(a, wires=0)
-            qml.RX(b, wires=1)
-            qml.CNOT(wires=[0, 1])
-            return [expval(qml.PauliZ(0)), expval(qml.PauliY(1))]
-
-        jac_fn = qml.jacobian(circuit)
-        res = jac_fn(a, b)
-        expected = [[-np.sin(a), 0], [np.sin(a) * np.sin(b), -np.cos(a) * np.cos(b)]]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        if diff_method == "finite-diff":
-            spy.assert_called()
-        elif diff_method == "backprop":
-            spy.assert_not_called()
-
-        # call the Jacobian with new parameters
-        a = np.array(0.6, requires_grad=True)
-        b = np.array(0.832, requires_grad=True)
-
-        res = jac_fn(a, b)
-        expected = [[-np.sin(a), 0], [np.sin(a) * np.sin(b), -np.cos(a) * np.cos(b)]]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-    def test_jacobian_options(self, dev_name, diff_method, mocker, tol):
-        """Test setting jacobian options"""
-        if diff_method == "backprop":
-            pytest.skip("Test does not support backprop")
-
-        spy = mocker.spy(QuantumTape, "numeric_pd")
-
-        a = np.array([0.1, 0.2], requires_grad=True)
-
-        dev = qml.device("default.qubit", wires=1)
-
-        @qnode(dev, interface="autograd", h=1e-8, order=2)
-        def circuit(a):
-            qml.RY(a[0], wires=0)
-            qml.RX(a[1], wires=0)
-            return expval(qml.PauliZ(0))
-
-        qml.jacobian(circuit)(a)
-
-        for args in spy.call_args_list:
-            assert args[1]["order"] == 2
-            assert args[1]["h"] == 1e-8
-
-    def test_changing_trainability(self, dev_name, diff_method, mocker, tol):
-        """Test changing the trainability of parameters changes the
-        number of differentiation requests made"""
-        if diff_method == "backprop":
-            pytest.skip("Test does not support backprop")
-
-        a = np.array(0.1, requires_grad=True)
-        b = np.array(0.2, requires_grad=True)
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qnode(dev, interface="autograd")
-        def circuit(a, b):
-            qml.RY(a, wires=0)
-            qml.RX(b, wires=1)
-            qml.CNOT(wires=[0, 1])
-            return expval(qml.PauliZ(0)), expval(qml.PauliY(1))
-
-        def loss(a, b):
-            return np.sum(circuit(a, b))
-
-        grad_fn = qml.grad(loss)
-
-        # TODO: uncomment when the parameter-shift rule is implemented
-        # spy = mocker.spy(QubitParamShiftTape, "_parameter_shift")
-        spy = mocker.spy(QuantumTape, "numeric_pd")
-
-        res = grad_fn(a, b)
-
-        # the tape has reported both arguments as trainable
-        assert circuit.qtape.trainable_params == {0, 1}
-
-        expected = [-np.sin(a) + np.sin(a) * np.sin(b), -np.cos(a) * np.cos(b)]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        # The parameter-shift rule has been called for each argument
-        assert len(spy.call_args_list) == 2
-
-        # make the second QNode argument a constant
-        a = np.array(0.54, requires_grad=True)
-        b = np.array(0.8, requires_grad=False)
-
-        spy.call_args_list = []
-        res = grad_fn(a, b)
-
-        # the tape has reported only the first argument as trainable
-        assert circuit.qtape.trainable_params == {0}
-
-        expected = [-np.sin(a) + np.sin(a) * np.sin(b)]
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        # QuantumTape.numeric_pd has been called only once
-        assert len(spy.call_args_list) == 1
-
-        # trainability also updates on evaluation
-        a = np.array(0.54, requires_grad=False)
-        b = np.array(0.8, requires_grad=True)
-        circuit(a, b)
-        assert circuit.qtape.trainable_params == {1}
-
-    def test_classical_processing(self, dev_name, diff_method, tol):
-        """Test classical processing within the quantum tape"""
-        a = np.array(0.1, requires_grad=True)
-        b = np.array(0.2, requires_grad=False)
-        c = np.array(0.3, requires_grad=True)
-
-        dev = qml.device(dev_name, wires=1)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(a, b, c):
-            qml.RY(a * c, wires=0)
-            qml.RZ(b, wires=0)
-            qml.RX(c + c ** 2 + np.sin(a), wires=0)
-            return expval(qml.PauliZ(0))
-
-        res = qml.jacobian(circuit)(a, b, c)
-
-        if diff_method == "finite-diff":
-            assert circuit.qtape.trainable_params == {0, 2}
-            tape_params = np.array(circuit.qtape.get_parameters())
-            assert np.all(tape_params == [a * c, c + c ** 2 + np.sin(a)])
-
-        assert res.shape == (1, 2)
-
-    def test_no_trainable_parameters(self, dev_name, diff_method, tol):
-        """Test evaluation and Jacobian if there are no trainable parameters"""
-        dev = qml.device(dev_name, wires=2)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(a, b):
-            qml.RY(a, wires=0)
-            qml.RX(b, wires=0)
-            qml.CNOT(wires=[0, 1])
-            return expval(qml.PauliZ(0)), expval(qml.PauliZ(1))
-
-        a = np.array(0.1, requires_grad=False)
-        b = np.array(0.2, requires_grad=False)
-
-        res = circuit(a, b)
-
-        if diff_method == "finite-diff":
-            assert circuit.qtape.trainable_params == set()
-
-        assert res.shape == (2,)
-        assert isinstance(res, np.ndarray)
-
-        assert not qml.jacobian(circuit)(a, b)
-
-        def cost(a, b):
-            return np.sum(circuit(a, b))
-
-        with pytest.warns(UserWarning, match="Output seems independent of input"):
-            grad = qml.grad(cost)(a, b)
-
-        assert grad == tuple()
-
-    def test_matrix_parameter(self, dev_name, diff_method, tol):
-        """Test that the autograd interface works correctly
-        with a matrix parameter"""
-        U = np.array([[0, 1], [1, 0]], requires_grad=False)
-        a = np.array(0.1, requires_grad=True)
-
-        dev = qml.device(dev_name, wires=2)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(U, a):
-            qml.QubitUnitary(U, wires=0)
-            qml.RY(a, wires=0)
-            return expval(qml.PauliZ(0))
-
-        res = circuit(U, a)
-
-        if diff_method == "finite-diff":
-            assert circuit.qtape.trainable_params == {1}
-
-        res = qml.grad(circuit)(U, a)
-        assert np.allclose(res, np.sin(a), atol=tol, rtol=0)
-
-    def test_differentiable_expand(self, dev_name, diff_method, mocker, tol):
-        """Test that operation and nested tapes expansion
-        is differentiable"""
-        mock = mocker.patch.object(qml.operation.Operation, "do_check_domain", False)
-
-        class U3(qml.U3):
-            def expand(self):
-                theta, phi, lam = self.data
-                wires = self.wires
-
-                with QuantumTape() as tape:
-                    qml.Rot(lam, theta, -lam, wires=wires)
-                    qml.PhaseShift(phi + lam, wires=wires)
-
-                return tape
-
-        dev = qml.device(dev_name, wires=1)
-        a = np.array(0.1, requires_grad=False)
-        p = np.array([0.1, 0.2, 0.3], requires_grad=True)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(a, p):
-            qml.RX(a, wires=0)
-            U3(p[0], p[1], p[2], wires=0)
-            return expval(qml.PauliX(0))
-
-        res = circuit(a, p)
-
-        if diff_method == "finite-diff":
-            assert circuit.qtape.trainable_params == {1, 2, 3, 4}
-        elif diff_method == "backprop":
-            # For a backprop device, no interface wrapping is performed, and QuantumTape.jacobian()
-            # is never called. As a result, QuantumTape.trainable_params is never set --- the ML
-            # framework uses its own backprop logic and its own bookkeeping re: trainable parameters.
-            assert circuit.qtape.trainable_params == {0, 1, 2, 3, 4}
-
-        assert [i.name for i in circuit.qtape.operations] == ["RX", "Rot", "PhaseShift"]
-
-        if diff_method == "finite-diff":
-            assert np.all(circuit.qtape.get_parameters() == [p[2], p[0], -p[2], p[1] + p[2]])
-        elif diff_method == "backprop":
-            # In backprop mode, all parameters are returned.
-            assert np.all(circuit.qtape.get_parameters() == [a, p[2], p[0], -p[2], p[1] + p[2]])
-
-        expected = np.cos(a) * np.cos(p[1]) * np.sin(p[0]) + np.sin(a) * (
-            np.cos(p[2]) * np.sin(p[1]) + np.cos(p[0]) * np.cos(p[1]) * np.sin(p[2])
-        )
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        res = qml.grad(circuit)(a, p)
-        expected = np.array(
-            [
-                np.cos(p[1]) * (np.cos(a) * np.cos(p[0]) - np.sin(a) * np.sin(p[0]) * np.sin(p[2])),
-                np.cos(p[1]) * np.cos(p[2]) * np.sin(a)
-                - np.sin(p[1])
-                * (np.cos(a) * np.sin(p[0]) + np.cos(p[0]) * np.sin(a) * np.sin(p[2])),
-                np.sin(a)
-                * (np.cos(p[0]) * np.cos(p[1]) * np.cos(p[2]) - np.sin(p[1]) * np.sin(p[2])),
-            ]
-        )
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-    def test_probability_differentiation(self, dev_name, diff_method, tol):
-        """Tests correct output shape and evaluation for a tape
-        with prob and expval outputs"""
-
-        dev = qml.device(dev_name, wires=2)
-        x = np.array(0.543, requires_grad=True)
-        y = np.array(-0.654, requires_grad=True)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return probs(wires=[0]), probs(wires=[1])
-
-        res = circuit(x, y)
-
-        expected = np.array(
-            [
-                [np.cos(x / 2) ** 2, np.sin(x / 2) ** 2],
-                [(1 + np.cos(x) * np.cos(y)) / 2, (1 - np.cos(x) * np.cos(y)) / 2],
-            ]
-        )
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        res = qml.jacobian(circuit)(x, y)
-        expected = np.array(
-            [
-                [[-np.sin(x) / 2, 0], [-np.sin(x) * np.cos(y) / 2, -np.cos(x) * np.sin(y) / 2]],
-                [
-                    [np.sin(x) / 2, 0],
-                    [np.cos(y) * np.sin(x) / 2, np.cos(x) * np.sin(y) / 2],
-                ],
-            ]
-        )
-
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-    def test_ragged_differentiation(self, dev_name, diff_method, monkeypatch, tol):
-        """Tests correct output shape and evaluation for a tape
-        with prob and expval outputs"""
-        dev = qml.device(dev_name, wires=2)
-        x = np.array(0.543, requires_grad=True)
-        y = np.array(-0.654, requires_grad=True)
-
-        if dev_name == "default.qubit.autograd":
-            # The current DefaultQubitAutograd device provides an _asarray method that does
-            # not work correctly for ragged arrays. For ragged arrays, we would like _asarray to
-            # flatten the array. Here, we patch the _asarray method on the device to achieve this
-            # behaviour; once the tape has moved from the beta folder, we should implement
-            # this change directly in the device.
-
-            def _asarray(args, dtype=np.float64):
-                return np.hstack(args).flatten()
-
-            monkeypatch.setattr(dev, "_asarray", _asarray)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit(x, y):
-            qml.RX(x, wires=[0])
-            qml.RY(y, wires=[1])
-            qml.CNOT(wires=[0, 1])
-            return [expval(qml.PauliZ(0)), probs(wires=[1])]
-
-        res = circuit(x, y)
-
-        expected = np.array(
-            [np.cos(x), (1 + np.cos(x) * np.cos(y)) / 2, (1 - np.cos(x) * np.cos(y)) / 2]
-        )
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-        res = qml.jacobian(circuit)(x, y)
-        expected = np.array(
-            [
-                [-np.sin(x), 0],
-                [-np.sin(x) * np.cos(y) / 2, -np.cos(x) * np.sin(y) / 2],
-                [np.cos(y) * np.sin(x) / 2, np.cos(x) * np.sin(y) / 2],
-            ]
-        )
-        assert np.allclose(res, expected, atol=tol, rtol=0)
-
-    def test_sampling(self, dev_name, diff_method):
-        """Test sampling works as expected"""
-        dev = qml.device(dev_name, wires=2, shots=10)
-
-        @qnode(dev, diff_method=diff_method, interface="autograd")
-        def circuit():
-            qml.Hadamard(wires=[0])
-            qml.CNOT(wires=[0, 1])
-            return [sample(qml.PauliZ(0)), sample(qml.PauliX(1))]
-
-        res = circuit()
-
-        assert res.shape == (2, 10)
-        assert isinstance(res, np.ndarray)
-
-
-def qtransform(qnode, a, interface=np):
-    """Transforms every RY(y) gate in a circuit to RX(-a*cos(y))"""
-
-    def construct(self, args, kwargs):
-        """New quantum tape construct method, that performs
-        the transform on the tape in a define-by-run manner"""
-
-        # the following global variable is defined simply for testing
-        # purposes, so that we can easily extract the transformed operations
-        # for verification.
-        global t_op
-
-        t_op = []
-
-        QNode.construct(self, args, kwargs)
-
-        new_ops = []
-        for o in self.qtape.operations:
-            # here, we loop through all tape operations, and make
-            # the transformation if a RY gate is encountered.
-            if isinstance(o, qml.RY):
-                t_op.append(qml.RX(-a * interface.cos(o.data[0]), wires=o.wires))
-                new_ops.append(t_op[-1])
-            else:
-                new_ops.append(o)
-
-        self.qtape._ops = new_ops
-        self.qtape._update()
-
-    import copy
-
-    new_qnode = copy.deepcopy(qnode)
-    new_qnode.construct = construct.__get__(new_qnode, QNode)
-    return new_qnode
-
-
-@pytest.mark.parametrize(
-    "dev_name,diff_method",
-    [("default.qubit", "finite-diff"), ("default.qubit.autograd", "backprop")],
-)
-def test_transform(dev_name, diff_method, monkeypatch, tol):
-    """Test an example transform"""
-    monkeypatch.setattr(qml.operation.Operation, "do_check_domain", False)
-
-    dev = qml.device(dev_name, wires=1)
-
-    @qnode(dev, interface="autograd", diff_method=diff_method)
-    def circuit(weights):
-        # the following global variables are defined simply for testing
-        # purposes, so that we can easily extract the operations for verification.
-        global op1, op2
-        op1 = qml.RY(weights[0], wires=0)
-        op2 = qml.RX(weights[1], wires=0)
-        return expval(qml.PauliZ(wires=0))
-
-    weights = np.array([0.32, 0.543], requires_grad=True)
-    a = np.array(0.5, requires_grad=True)
-
-    def loss(weights, a):
-        # the following global variable is defined simply for testing
-        # purposes, so that we can easily extract the transformed QNode
-        # for verification.
-        global new_circuit
-
-        # transform the circuit QNode with trainable weight 'a'
-        new_circuit = qtransform(circuit, a)
-
-        # evaluate the transformed QNode
-        res = new_circuit(weights)[0]
-
-        # evaluate the original QNode with pre-processed parameters
-        res2 = circuit(np.sin(weights))[0]
-
-        # return the sum of the two QNode evaluations
-        return res + res2
-
-    res = loss(weights, a)
-
-    # verify that the transformed QNode has the expected operations
-    assert circuit.qtape.operations == [op1, op2]
-    assert new_circuit.qtape.operations[0] == t_op[0]
-    assert new_circuit.qtape.operations[1].name == op2.name
-    assert new_circuit.qtape.operations[1].wires == op2.wires
-
-    # check that the incident gate arguments of both QNode tapes are correct
-    assert np.all(np.array(circuit.qtape.get_parameters()) == np.sin(weights))
-    assert np.all(
-        np.array(new_circuit.qtape.get_parameters()) == [-a * np.cos(weights[0]), weights[1]]
-    )
-
-    # verify that the gradient has the correct shape
-    grad = qml.grad(loss)(weights, a)
-    assert len(grad) == 2
-    assert grad[0].shape == weights.shape
-    assert grad[1].shape == a.shape
-
-    # compare against the expected values
-    assert np.allclose(res, 1.8244501889992706, atol=tol, rtol=0)
-    assert np.allclose(grad[0], [-0.26610258, -0.47053553], atol=tol, rtol=0)
-    assert np.allclose(grad[1], 0.06486032, atol=tol, rtol=0)
